refactor(json): split chart semantics from pure IR
