+++ packages/chart/package.json
@@ -1,10 +1,7 @@
 {
-  "name": "@be-music/player",
+  "name": "@be-music/chart",
   "version": "0.1.0",
-  "description": "CLI BMS player",
-  "bin": {
-    "bms-player": "dist/cli.js"
-  },
+  "description": "Chart semantics and playback helpers for Be-Music IR",
   "files": [
     "dist"
   ],
@@ -21,9 +18,7 @@
   "scripts": {
     "build:bundle": "vite build",
     "build": "vite build && tsgo -b tsconfig.json",
-    "build:sea": "tsx ../../scripts/build-sea.ts --package player",
     "clean": "rimraf dist tsconfig.tsbuildinfo",
-    "dev": "tsx --tsconfig ../../tsconfig.typecheck.json src/cli.ts",
     "typecheck": "tsgo -p tsconfig.typecheck.json --pretty",
     "lint": "oxlint .",
     "lint:fix": "oxlint . --fix",
@@ -32,19 +27,6 @@
     "bench": "tsx --tsconfig ../../tsconfig.typecheck.json scripts/bench.ts"
   },
   "dependencies": {
-    "@be-music/audio-renderer": "workspace:*",
-    "@be-music/chart": "workspace:*",
-    "@be-music/json": "workspace:*",
-    "@be-music/parser": "workspace:*",
-    "@be-music/utils": "workspace:*",
-    "@uwx/libav.js-fat": "6.0.0-nightly.29.f420ff.ffmpeg.6.1.1",
-    "alien-signals": "^3.1.2",
-    "fast-bmp": "^4.0.1",
-    "fast-png": "^8.0.0",
-    "jpeg-js": "^0.4.4",
-    "node-web-audio-api": "^1.0.8"
-  },
-  "devDependencies": {
-    "postject": "^1.0.0-alpha.6"
+    "@be-music/json": "workspace:*"
   }
 }
