+++ packages/editor/package.json
@@ -31,6 +31,7 @@
     "bench": "tsx --tsconfig ../../tsconfig.typecheck.json scripts/bench.ts"
   },
   "dependencies": {
+    "@be-music/chart": "workspace:*",
     "@be-music/json": "workspace:*",
     "@be-music/parser": "workspace:*",
     "@be-music/stringifier": "workspace:*",
