+++ packages/json/scripts/exports-cases.ts
@@ -27,125 +27,10 @@ export function registerJsonExportsCases(define: DefineBenchmarkCase): void {
       jsonApi.intToBase36(12_345, 2);
     },
   });
-  define('json.parseBpmFrom03Token', {
-    run: () => {
-      jsonApi.parseBpmFrom03Token('7F');
-    },
-  });
   define('json.ensureMeasure', {
     run: (fixtures) => {
       const cloned = jsonApi.cloneJson(fixtures.sampleBmsJson);
       jsonApi.ensureMeasure(cloned, 32);
-    },
-  });
-  define('json.getMeasureBeats', {
-    run: () => {
-      jsonApi.getMeasureBeats(1.5);
-    },
-  });
-  define('json.measureToBeat', {
-    run: (fixtures) => {
-      jsonApi.measureToBeat(fixtures.sampleBmsJson, 2, 0.5);
-    },
-  });
-  define('json.eventToBeat', {
-    run: (fixtures) => {
-      jsonApi.eventToBeat(fixtures.sampleBmsJson, fixtures.eventA);
-    },
-  });
-  define('json.createBeatResolver', {
-    run: (fixtures) => {
-      jsonApi.createBeatResolver(fixtures.sampleBmsJson);
-    },
-  });
-  define('json.sortEvents', {
-    run: (fixtures) => {
-      jsonApi.sortEvents(fixtures.sampleBmsJson.events);
-    },
-  });
-  define('json.compareEvents', {
-    run: (fixtures) => {
-      jsonApi.compareEvents(fixtures.eventA, fixtures.eventB);
-    },
-  });
-  define('json.isTempoChannel', {
-    run: () => {
-      jsonApi.isTempoChannel('08');
-    },
-  });
-  define('json.isStopChannel', {
-    run: () => {
-      jsonApi.isStopChannel('09');
-    },
-  });
-  define('json.isScrollChannel', {
-    run: () => {
-      jsonApi.isScrollChannel('SC');
-    },
-  });
-  define('json.isLandmineChannel', {
-    run: () => {
-      jsonApi.isLandmineChannel('D1');
-    },
-  });
-  define('json.isSampleTriggerChannel', {
-    run: () => {
-      jsonApi.isSampleTriggerChannel('11');
-    },
-  });
-  define('json.isPlayableChannel', {
-    run: () => {
-      jsonApi.isPlayableChannel('11');
-    },
-  });
-  define('json.isPlayLaneSoundChannel', {
-    run: () => {
-      jsonApi.isPlayLaneSoundChannel('31');
-    },
-  });
-  define('json.mapBmsLongNoteChannelToPlayable', {
-    run: () => {
-      jsonApi.mapBmsLongNoteChannelToPlayable('51');
-    },
-  });
-  define('json.isBmsLongNoteChannel', {
-    run: () => {
-      jsonApi.isBmsLongNoteChannel('61');
-    },
-  });
-  define('json.isBmsBgmVolumeChangeChannel', {
-    run: () => {
-      jsonApi.isBmsBgmVolumeChangeChannel('97');
-    },
-  });
-  define('json.isBmsKeyVolumeChangeChannel', {
-    run: () => {
-      jsonApi.isBmsKeyVolumeChangeChannel('98');
-    },
-  });
-  define('json.isBmsDynamicVolumeChangeChannel', {
-    run: () => {
-      jsonApi.isBmsDynamicVolumeChangeChannel('97');
-    },
-  });
-  define('json.parseBmsDynamicVolumeGain', {
-    run: () => {
-      jsonApi.parseBmsDynamicVolumeGain('80');
-    },
-  });
-  define('json.resolveBmsLongNotes', {
-    run: (fixtures) => {
-      jsonApi.resolveBmsLongNotes(fixtures.sampleBmsJson, { inferLnTypeWhenMissing: true });
-    },
-  });
-  define('json.collectLnobjEndEvents', {
-    run: (fixtures) => {
-      jsonApi.collectLnobjEndEvents(fixtures.sampleBmsJson);
-    },
-  });
-  define('json.resolveLnobjLongNotes', {
-    run: (fixtures) => {
-      jsonApi.resolveLnobjLongNotes(fixtures.sampleBmsJson);
     },
   });
 }
