+++ packages/parser/tsconfig.json
@@ -8,5 +8,5 @@
   },
   "include": ["src"],
   "exclude": ["src/**/*.test.ts", "src/**/*.spec.ts"],
-  "references": [{ "path": "../json" }, { "path": "../utils" }]
+  "references": [{ "path": "../json" }, { "path": "../chart" }, { "path": "../utils" }]
 }
