+++ packages/player/src/core/timeline.test.ts
@@ -1,5 +1,6 @@
 import { createTimingResolver } from '@be-music/audio-renderer';
-import { createBeatResolver, createEmptyJson, type BeMusicEvent } from '@be-music/json';
+import { createBeatResolver } from '@be-music/chart';
+import { createEmptyJson, type BeMusicEvent } from '@be-music/json';
 import { describe, expect, test } from 'vitest';
 import {
   createBeatAtSecondsResolver,
