+++ scripts/bench/packages/chart.ts
@@ -1,0 +1,8 @@
+import * as chartApi from '@be-music/chart';
+import { registerChartExportsCases } from '../../../packages/chart/scripts/exports-cases.ts';
+import type { BenchmarkPackageDefinition } from '../exports.types.ts';
+
+export const chartBenchmarkPackage: BenchmarkPackageDefinition = {
+  module: chartApi as Record<string, unknown>,
+  registerCases: registerChartExportsCases,
+};
