+++ scripts/bench/packages/index.ts
@@ -5,6 +5,7 @@ import {
   type PackageName,
 } from '../exports.types.ts';
 import { audioRendererBenchmarkPackage } from './audio-renderer.ts';
+import { chartBenchmarkPackage } from './chart.ts';
 import { editorBenchmarkPackage } from './editor.ts';
 import { jsonBenchmarkPackage } from './json.ts';
 import { parserBenchmarkPackage } from './parser.ts';
@@ -15,6 +16,7 @@ import { utilsBenchmarkPackage } from './utils.ts';
 export const PACKAGE_DEFINITIONS: Record<PackageName, BenchmarkPackageDefinition> = {
   utils: utilsBenchmarkPackage,
   json: jsonBenchmarkPackage,
+  chart: chartBenchmarkPackage,
   parser: parserBenchmarkPackage,
   stringifier: stringifierBenchmarkPackage,
   editor: editorBenchmarkPackage,
